Move some files around.
