+++ experiments/billable/src/app/pages/invoice/DetailPage/InvoiceDetailPage.tsx
@@ -3,7 +3,7 @@
 import { Layout } from "../../Layout";
 
 import { InvoiceForm } from "./InvoiceForm";
-import { RouteContext } from "../../../../router";
+import { RouteContext } from "../../../../lib/router";
 import { db } from "../../../../db";
 
 export type InvoiceItem = {
